# server/docs/swagger.yaml
basePath: /api
definitions:
  analitics.byDistrictStatistics:
    properties:
      alive:
        description: количество живых коров
        type: integer
      districtID:
        description: ID района
        type: integer
      genotyped:
        description: количество генотипированных коров
        type: integer
      ill:
        description: количество больных коров
        type: integer
    type: object
...
  analitics.byHozStatistics:
    properties:
      alive:
        description: количество живых коров
        type: integer
      genotyped:
        description: количество генотипированных коров
        type: integer
      ill:
        description: количество больных коров
        type: integer
    type: object
...
  auth.AuthData:
    properties:
      email:
        description: почта
        type: string
      password:
        description: пароль, не зашифрованный
        type: string
    type: object
...
  cows_filter.CowsFilter:
    properties:
      birkingDateFrom:
        description: Фильтр по дате перебирковки коровы, ищет коров у которых быа
          перебирковка в эту дату или позднее
        example: "1800-01-21"
        type: string
      birkingDateTo:
        description: Фильтр по дате осеменения коровы, ищет коров у которых была перебирковка
          в эту дату или позднее
        example: "2800-01-21"
        type: string
      birthDateFrom:
        description: RegionId               *uint   `example:"1" validate:"optional"`          //ID
          региона, для которого ищутся коровы
        example: "1800-01-21"
        type: string
      birthDateTo:
        description: Фильтр по дню рождения коровы ОТ (возращает всех кто родился
          в эту дату или раньше)
        example: "2800-01-21"
        type: string
      breedId:
        description: Фильтр по ID пород несколько ID пород - возращает всех коров,
          ID пород которых в списке
        items:
          type: integer
        type: array
      calvingDateFrom:
        description: Фильтр по дате отела коровы, ищет коров у которых был отел в
          эту дату или позднее
        example: "1800-01-21"
        type: string
      calvingDateTo:
        description: Фильтр по дате осеменения коровы, ищет коров у которых был отел
          в эту дату или позднее
        example: "2800-01-21"
        type: string
      controlMilkingDateFrom:
        description: Фильтр по дате контрольной дойки, ищет коров у которых была контрольная
          дойка в эту дату или позднее
        example: "1800-01-21"
        type: string
      controlMilkingDateTo:
        description: Фильтр по дате контрольной дойки, ищет коров у которых была контрольная
          дойка в эту дату или ранее
        example: "2800-01-21"
        type: string
      createdAtFrom:
        description: Фильтр по дате внесения записи в БД ОТ
        type: string
      createdAtTo:
        description: Фильтр по дате внесения записи в БД ДО
        type: string
      departDateFrom:
        description: Фильтр по дате открепления коровы ищет всех коров открепленных
          от коровника в эту дату или позднее
        example: "1800-01-21"
        type: string
      departDateTo:
        description: Фильтр по дате открепления коровы ищет всех коров открепленных
          от коровника в эту дату или раньше
        example: "2800-01-21"
        type: string
      entitiesOnPage:
        default: 50
        description: Количество сущностей на странице
        type: integer
      exteriorFrom:
        default: 3.14
        description: Фильтр по оценке экстерьера коровы ОТ
        type: number
      exteriorTo:
        default: 3.14
        description: Фильтр по оценке экстерьера коровы ДО
        type: number
      genotypingDateFrom:
        description: фильтр по дате генотипирования ОТ
        example: "1800-01-21"
        type: string
      genotypingDateTo:
        description: фильтр по дате генотипирования ДО
        example: "2800-01-21"
        type: string
      hasAnyIllnes:
        default: false
        description: Флаг true - возращает коров у которых есть хотябы одно заболевение,
          false - возращает коров, у которых нет ни одного
        type: boolean
      hozId:
        description: ID фермы, для которой ищутся коровы
        example: 1
        type: integer
      illDateFrom:
        description: Фильтр по дате начала болезни ОТ
        example: "1800-01-21"
        type: string
      illDateTo:
        description: Фильтр по дате начала болезни ДО
        example: "1800-01-21"
        type: string
      inbrindingCoeffByFamilyFrom:
        default: 3.14
        description: фильтр по коэф. инбриндинга по роду ОТ
        type: number
      inbrindingCoeffByFamilyTo:
        default: 3.14
        description: фильтр по коэф. инбриндинга по роду ДО
        type: number
      inbrindingCoeffByGenotypeFrom:
        default: 3.14
        description: фильтр по коэф. инбриндинга по генотипу ОТ
        type: number
      inbrindingCoeffByGenotypeTo:
        default: 3.14
        description: фильтр по коэф. инбриндинга по генотипу ДО
        type: number
      inseminationDateFrom:
        description: Exterior             *float64 `default:"3.14" validate:"optional"`       //
          Фильтр по оценке экстерьера коровы, будет переработан
        example: "1800-01-21"
        type: string
      inseminationDateTo:
        description: Фильтр по дате осеменения коровы, ищет коров у которых было осеменение
          в эту дату или ранее
        example: "2800-01-21"
        type: string
      isAborted:
        default: false
        description: Фильтр по абортам Был/ не был
        type: boolean
      isDead:
        default: false
        description: Фильтр живых/мертвых коров (true - ищет мертвых, false - живых)
        type: boolean
      isGenotyped:
        description: Фильтр по наличию информации о генотипировании
        type: boolean
      isIll:
        default: false
        description: Фильтр по наличию генетических заболеваний
        type: boolean
      isStillBorn:
        default: false
        description: Фильтр по мертворождению Было/не было
        type: boolean
      isTwins:
        default: false
        description: Фильтр по родам двойняшек Было / не было
        type: boolean
      monogeneticIllneses:
        description: ID ген. заболеваний из /api/mongenetic_illnesses
        items:
          type: integer
        type: array
      orderBy:
        description: 'Может принимать следующие и только следующие значения: null,
          "RSHN", "InventoryNumber", "Name", "HozName", "BirthDate"'
        type: string
      orderByDesc:
        description: true - в порядке убывания, false - в порядке возрастания
        type: boolean
      pageNumber:
        default: 1
        description: Номер страницы для отображения
        type: integer
      regionId:
        description: по региону
        type: integer
      searchQuery:
        description: Имя, номер РСХН или инвентарный номер, по которым ищется корова
        example: Буренка
        type: string
      sex:
        description: ID пола коровы (если нужно несколько разных полов - несколько
          ID)
        items:
          type: integer
        type: array
    type: object
...
host: genmilk.ru
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Сваггер сгенерирован из кода, поэтому может содержать неточности. По мере возможности они будут описаны далее
    Все даты передаются как строки
    Большая часть рутов не возвращает вложенные объекты
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GenMilk API
  version: "1.0"
paths:
...
  /analitics/genotyped/{year}/byDistrict/{district}/hoz:
    post:
      description: Возращает словарь хозяйство - количество живых коров, количество
        генотипированных
      parameters:
      - description: год за который собирается статистика
        in: path
        name: year
        required: true
        type: integer
      - description: район за который собирается статистика
        in: path
        name: district
        required: true
        type: integer
      - description: applied filters
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/cows_filter.CowsFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties:
                $ref: '#/definitions/analitics.byHozStatistics'
              type: object
            type: array
        "421":
          description: Misdirected Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get analytics by hoz
      tags:
      - Analytics(GeneticFilters)
...
  /analitics/genotyped/{year}/byRegion/{region}/districts:
    get:
      description: Возращает словарь район - количество живых коров, количество генотипированных
      parameters:
      - description: год за который собирается статистика
        in: path
        name: year
        required: true
        type: integer
      - description: регион за который собирается статистика
        in: path
        name: region
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/analitics.byDistrictStatistics'
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Get genetic analytics by districts
      tags:
      - Analytics (GeneticGET)
    post:
      description: Возращает словарь район - количество живых коров, количество генотипированных
      parameters:
      - description: год за который собирается статистика
        in: path
        name: year
        required: true
        type: integer
      - description: регион за который собирается статистика
        in: path
        name: region
        required: true
        type: integer
      - description: applied filters
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/cows_filter.CowsFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties:
                $ref: '#/definitions/analitics.byDistrictStatistics'
              type: object
            type: array
        "421":
          description: Misdirected Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get analytics by districts
      tags:
      - Analytics(GeneticFilters)
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: После успешного логина возвращает словарь с ключем "token" - access
        token. Обеспечивает логин по JWT
      parameters:
      - description: applied filters
        in: body
        name: AuthData
        required: true
        schema:
          $ref: '#/definitions/auth.AuthData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: {}
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: {}
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: {}
            type: object
      summary: LOGIN
      tags:
      - LOGIN
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"